Basic Forest Generation Tool
First Implementation of Foliage Generation

Successes:
Now able to generate trees using noise map heights and raycast.

Still needs to be done:
The tree generation only works when mesh is of size 1.

Still need to attach the heights of the textures(grass, sand, water) to the generation heights, as right now it is just set in script to 0.4.